fixed user.uid issue delete workout function

+++ src/Workout.js
@@ -1,27 +1,26 @@
 import React from 'react';
 import { useState, useEffect } from 'react';
 import { db } from "./firebase-config";
-import { collection, getDocs, doc, deleteDoc } from "firebase/firestore";
+import { collection, getDocs, doc, updateDoc, deleteField, deleteDoc, arrayRemove, FieldValue } from "firebase/firestore";
 
 const Workout = ({ user }) => {
     
     const [ allWorkouts, setAllWorkouts ] = useState([]);
 
-    const [ newExercise, setNewExercise ] = useState("");
-    const [ newWeight, setNewWeight ] = useState(0);
-    const [ newReps, setNewReps ] = useState(0);
+    // const [ newExercise, setNewExercise ] = useState("");
+    // const [ newWeight, setNewWeight ] = useState(0);
+    // const [ newReps, setNewReps ] = useState(0);
 
     // get all workouts
     useEffect(() => {
       const getWorkouts = async () => {
         try {
           const snapshot = await getDocs(
-            // must alter to take user.uid
-            collection(db, "users", "hwUQMGclPZQCJUO1jgYPDnzQwZp2", "workouts")
+            collection(db, "users", `${user.uid}`, "workouts")
           );
           let tempArr = [];
           snapshot.forEach((doc) => {
-            tempArr.push({ ...doc.data() });
+            tempArr.push({ ...doc.data(), id: doc.id });
           })
           setAllWorkouts([...tempArr])
         } catch (error) {
@@ -29,45 +28,43 @@ const Workout = ({ user }) => {
         }
       }
       getWorkouts();
-    }, []);
+    }, [user]);
 
-    // need to sort doc id
-    const editWorkout = async (id) => {
-      console.log(id)
-    }
-
-    // need to alter to accept workout id
     const deleteWorkout = async (id) => {
-      console.log(id)
       try {    
-        await deleteDoc(
-          doc(
-            db,
-            "users",
-            "hwUQMGclPZQCJUO1jgYPDnzQwZp2",
-            "workouts",
-            id
-          )
-        );
-        console.log('deleted')
+        await deleteDoc(doc(db, "users", `${user.uid}`, "workouts", id));
+        console.log('deleted workout')
       } catch (error) {
         console.log(error.message)
       }
-      
     }
 
-    // console.log(user.uid)
+    // not working
+    const deleteExercise = async (index, id) => {
+      // try {
+      //   const docRef = doc(db, "users", `${user.uid}`, "workouts", id);
+      //   await updateDoc(docRef, { exercises: arrayRemove(index)});
+      // } catch (error) {
+      //   console.log(error.message);
+      // }
+    }
+
+    const editExercise = async (index) => {
+      // display modal with input feilds
+      // set states to user inputs
+      // delete array item being edited
+      // push new item to array using states
+    };
 
   return (
     <>
       <h1>Read Workout Component</h1>
       <table>          
           {
-            allWorkouts.map((workout, index) => {
+            allWorkouts.map((workout) => {
               return (
-                <div key={index}>
-                  <button onClick={() => {deleteWorkout(index)}}>Delete workout</button>
-                  <button onClick={() => {editWorkout(index)}}>Edit Workout</button>
+                <div key={workout.id}>
+                  <button onClick={() => {deleteWorkout(workout.id)}}>Delete workout</button>
                   <tr>
                     <th rowSpan={2}>Exercise</th>
                     <th colSpan={3}>
@@ -80,14 +77,17 @@ const Workout = ({ user }) => {
                     <th>Reps</th>
                     <th>Total</th>
                     <th>Progress?</th>
+                    <th colSpan={2}>Options</th>
                   </tr>
-                  {workout.exercises.map((exercise) => (
-                    <tr>
+                  {workout.exercises.map((exercise, index) => (
+                    <tr key={index}>
                       <td>{exercise?.exercise}</td>
                       <td>{exercise?.weight}</td>
                       <td>{exercise?.reps}</td>
                       <td>{exercise.weight * exercise.reps}</td>
                       <td>Yes/No</td>
+                      <td><button onClick={() => {deleteExercise(index, workout.id)}}>Delete</button></td>
+                      <td><button onClick={() => {editExercise(index)}}>Edit</button></td>
                     </tr>
                   ))}
                 </div>
